<Col>: move col manipulation completely to ColRow
